Move `accumulate_result_with` helper function into `prunable` submodule
