+++ shardtree/src/lib.rs
@@ -1475,24 +1475,6 @@ impl<
     }
 }
 
-// We need an applicative functor for Result for this function so that we can correctly
-// accumulate errors, but we don't have one so we just write a special- cased version here.
-fn accumulate_result_with<A, B, C>(
-    left: Result<A, Vec<Address>>,
-    right: Result<B, Vec<Address>>,
-    combine_success: impl FnOnce(A, B) -> C,
-) -> Result<C, Vec<Address>> {
-    match (left, right) {
-        (Ok(a), Ok(b)) => Ok(combine_success(a, b)),
-        (Err(mut xs), Err(mut ys)) => {
-            xs.append(&mut ys);
-            Err(xs)
-        }
-        (Ok(_), Err(xs)) => Err(xs),
-        (Err(xs), Ok(_)) => Err(xs),
-    }
-}
-
 #[cfg(test)]
 mod tests {
     use assert_matches::assert_matches;
